api/requirements.txt
Flask
pandas
scipy
numpy
psycopg2
gunicorn
pmdarima
statsmodels
scikit-learn
matplotlib
seaborn
google-cloud-storage
fsspec
gcsfs